+++ nwb_web_gui/dashapps/pages/converter/example_schemas/metada_schema_2.json
