+++ nwb_web_gui/dashapps/pages/converter/example_schemas/metadata_example.json
