+++ nwb_web_gui/dashapps/pages/converter/example_schemas/source_schema.json
